medgemma-chest-xray-analyzer/medgemma-chest-xray-analyzer/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.50.0
accelerate>=0.26.0
bitsandbytes>=0.42.0
Pillow>=10.0.0
gradio>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
